server/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
google-genai
python-dotenv
